provide CRUD API as a sub-package of Gin
